modified stdout messages
Warnings for undefined keys are only issued if the key is "None" in
the 'union' dictionary AND it is undefined in some non-default mode.

keys do not have to appear in the default mode if they are fully
defined elsewhere.

When the modes are appended to the acf file, a message returns
reporting that to be the case.